Rename analyze_jar to analyze_container.
