+++ extlib/millhone/src/main.rs
@@ -112,7 +112,7 @@ enum Commands {
     Commit(cmd::commit::Subcommand),
 
     /// Find and fingerprint JAR files.
-    AnalyzeJars(cmd::analyze_jars::Subcommand),
+    AnalyzeContainer(cmd::analyze_container::Subcommand),
 }
 
 fn main() -> stable_eyre::Result<()> {
@@ -155,7 +155,7 @@ fn main() -> stable_eyre::Result<()> {
         Commands::Ingest(opts) => cmd::ingest::main(&app.direct_endpoint, opts),
         Commands::Analyze(opts) => cmd::analyze::main(&app.direct_endpoint, opts),
         Commands::Commit(opts) => cmd::commit::main(opts),
-        Commands::AnalyzeJars(opts) => cmd::analyze_jars::main(opts),
+        Commands::AnalyzeContainer(opts) => cmd::analyze_container::main(opts),
     }
 }
